fixed some bugs and changed the eyetracking request so it will try to use only the data for one eye if its possible

+++ server/MA-FF-Server/MA-FF-Server/Server/ConnectionManager.cs
@@ -79,7 +79,7 @@ namespace WebAnalyzer.Server
             this.Broadcast(message);
         }
 
-        public void RequestData(String uniqueId, int xPos, int yPos)
+        public void RequestData(String uniqueId, double xPos, double yPos)
         {
             SmallDataMessage message = new SmallDataMessage(Timestamp.GetMillisecondsUnixTimestamp());
